+++ module/actor/actor.js
@@ -109,6 +109,10 @@ export class Actor4e extends Actor {
 		for (let [id, abl] of Object.entries(data.abilities)) {
 
 			abl.mod = Math.floor((abl.value - 10) / 2);
+			if(abl.mod){
+				this.update({[`data.abilities.${id}.mod`]: abl.mod});
+			}
+
 			abl.modHalf = abl.mod + Math.floor(data.details.level / 2);
 			abl.prof = (abl.proficient || 0);
 			abl.saveBonus = saveBonus + Math.floor(data.details.level / 2);
@@ -159,7 +163,7 @@ export class Actor4e extends Actor {
 		}
 		
 		data.details.bloodied = Math.floor(data.attributes.hp.max / 2);
-		data.details.surgeValue = Math.floor(data.details.bloodied / 2) + data.details.surgeBon.value;
+		data.details.surgeValue += Math.floor(data.details.bloodied / 2) + data.details.surgeBon.value;
 		data.attributes.hp.min = -data.details.bloodied;
 		data.details.secondWindValue = data.details.surgeValue + data.details.secondwindbon.value;
 
@@ -208,8 +212,8 @@ export class Actor4e extends Actor {
 			}
 		}
 		
-		let tier = Math.clamped(Math.floor(( data.details.level - 1 ) /10 + 1),1,3);
-		this.update({[`data.details.tier`]: tier });
+		data.details.tier = Math.clamped(Math.floor(( data.details.level - 1 ) /10 + 1),1,3);
+		this.update({[`data.details.tier`]: data.details.tier });
 
 		//Weight & Encumbrance
 		data.encumbrance = this._computeEncumbrance(actorData);